create intermediary script for procurement preparation
