+++ rules/solve_myopic.smk
@@ -113,6 +113,40 @@ rule add_brownfield:
 ruleorder: add_existing_baseyear > add_brownfield
 
 
+rule add_procurement:
+    params:
+        procurement_enable=config_provider("enable", "procurement"),
+        ci_load=config_provider("enable", "ci_load"),
+        electricity=config_provider("electricity"),
+        procurement=config_provider("procurement"),
+        costs=config_provider("costs"),
+        max_hours=config_provider("electricity", "max_hours"),
+    input:
+        network=resources(
+            "networks/base_s_{clusters}_{opts}_{sector_opts}_{planning_horizons}_brownfield.nc"
+        ),
+        costs=resources("costs_{planning_horizons}.csv"),
+    output:    
+        network=resources(
+            "networks/base_s_{clusters}_{opts}_{sector_opts}_{planning_horizons}_procurement.nc"
+        ),
+    conda:
+        "../envs/environment.yaml"
+    script:
+        "../scripts/add_procurement.py"
+
+
+def choose_network_file(wildcards):
+    base_template = "networks/base_s_{clusters}_{opts}_{sector_opts}_{planning_horizons}"
+    
+    if config_provider("enable", "procurement") or config_provider("enable", "ci_load"):
+        suffix = "_procurement.nc"
+    else:
+        suffix = "_brownfield.nc"
+    
+    return resources(base_template + suffix)
+
+
 rule solve_sector_network_myopic:
     params:
         solving=config_provider("solving"),
@@ -120,20 +154,9 @@ rule solve_sector_network_myopic:
         co2_sequestration_potential=config_provider(
             "sector", "co2_sequestration_potential", default=200
         ),
-        # ================ New Params Start ================
-        procurement_enable=config_provider("enable", "procurement"),
-        ci_load=config_provider("enable", "ci_load"),
-        electricity=config_provider("electricity"),
-        procurement=config_provider("procurement"),
-        costs=config_provider("costs"),
-        max_hours=config_provider("electricity", "max_hours"),
-        # ================  New Params End  ================
         custom_extra_functionality=input_custom_extra_functionality,
     input:
-        network=resources(
-            "networks/base_s_{clusters}_{opts}_{sector_opts}_{planning_horizons}_brownfield.nc"
-        ),
-        costs=resources("costs_{planning_horizons}.csv"),
+        network = lambda w: choose_network_file(w)
     output:
         network=RESULTS
         + "networks/base_s_{clusters}_{opts}_{sector_opts}_{planning_horizons}.nc",
